feat(question): TikZ question with PLTikZ module
This is the same question that was done for PR #15, but this time using
a module called PLTikZ to streamline the process. PLTikZ deals with the
bulk of the TikZ-related work, and TikZ questions can now be
parameterized similarly to PL questions using Jinja. Some stuff may
still change with PLTikZ, but this is a great start!